requirements.txt
django >= 4.0.0 , <4.1.2
djangorestframework >= 3.14.0
django-environ==0.9.0
django-extensions
django-filter==22.1
psycopg2-binary==2.9.3
djangorestframework-simplejwt >= 5.2.2
pyjwt >= 2.6.0
# Pillow >= 8.4 ,<= 9.2.0